+++ src/DTS/eBaySDK/ResolutionCaseManagement/Services/ResolutionCaseManagementBaseService.php
@@ -54,7 +54,7 @@ class ResolutionCaseManagementBaseService extends \DTS\eBaySDK\Services\BaseServ
             );
         }
 
-        parent::__construct('http://svcs.ebay.com/services/resolution/v1/ResolutionCaseManagementService', 'https://svcs.sandbox.ebay.com/services/resolutionn/v1/ResolutionCaseManagement', $config, $httpClient);
+        parent::__construct('https://svcs.ebay.com/services/resolution/v1/ResolutionCaseManagementService', 'https://svcs.sandbox.ebay.com/services/resolution/v1/ResolutionCaseManagementService', $config, $httpClient);
     }
 
     /**